# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Payroll",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/payroll-periods": {
            "get": {
                "tags": [
                    "Payroll Periods"
                ],
                "summary": "Retrieve payroll periods with optional filtering",
                "operationId": "870dba619abc180810c585759c85e3d0",
                "parameters": [
                    {
                        "name": "has_filter",
                        "in": "query",
                        "description": "Set to false to get all payroll periods",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true
                        }
                    },
                    {
                        "name": "employment_type",
                        "in": "query",
                        "description": "Filter by employment type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "period_type",
                        "in": "query",
                        "description": "Filter by period type",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "description": "Filter by month (1-12)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 12,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Filter by year (e.g., 2025)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payroll periods retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payroll-periods/{id}": {
            "put": {
                "tags": [
                    "Payroll Periods"
                ],
                "summary": "Set or lock payroll period",
                "operationId": "6d4a6136e1ac8c68ca72d41f77523abb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Payroll period ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "set_period"
                                ],
                                "properties": {
                                    "set_period": {
                                        "description": "Set as active period",
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payroll period updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Payroll period not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}